+++ src/main/java/es/mdef/gestionusuarios/repositorios/UsuarioRepositorio.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import es.mdef.gestionusuarios.entidades.Usuario;
 
-public interface UsuarioRepositorio extends JpaRepository<Usuario, Integer> {
-	    
+public interface UsuarioRepositorio extends JpaRepository<Usuario, Long> {	    
         List<Usuario> findUsuarioByNombre(String nombre);
 }
